myapp/serializers.py
from rest_framework import serializers
from .models import PullRequest

class PullRequestSerializer(serializers.ModelSerializer):
    class Meta:
        model = PullRequest
        fields = '__all__'
