requirements.txt
asgiref==3.6.0
async-generator==1.10
attrs==23.1.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
colorama==0.4.6
Django==4.2.1
django-rest-framework==0.1.0
djangorestframework==3.14.0
exceptiongroup==1.1.1
h11==0.14.0
idna==3.4
outcome==1.2.0
packaging==23.1
Pillow==9.5.0
pycparser==2.21
PySocks==1.7.1
python-dotenv==1.0.0
pytz==2023.3
requests==2.30.0
selenium==4.9.1
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.4
tqdm==4.65.0
trio==0.22.0
trio-websocket==0.10.2
tzdata==2023.3
urllib3==2.0.2
webdriver-manager==3.8.6
wsproto==1.2.0
